% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcgis_extras.R
\name{arc_rast}
\alias{arc_rast}
\title{read spatraster using arcgisbinding}
\usage{
arc_rast(dsn)
}
\arguments{
\item{dsn}{The path of the source file}
}
\value{
object of class \code{SpatRaster} when a file was successfully read.
}
\description{
read spatraster using arcgisbinding
}
\examples{
\dontrun{
gdbpath = system.file('extdata/xian.gdb',package = 'georhythm')
elev = arc_rast(paste0(gdbpath,'/elev'))
}
}
