% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcgis_extras.R
\name{arc_read}
\alias{arc_read}
\title{read simple features using arcgisbinding}
\usage{
arc_read(dsn)
}
\arguments{
\item{dsn}{The path of the source file}
}
\value{
object of class \code{sf} when a file was successfully read.
}
\description{
read simple features using arcgisbinding
}
\examples{
\dontrun{
gdbpath = system.file('extdata/xian.gdb',package = 'georhythm')
block = arc_read(paste0(gdbpath,'/urban'))
}
}
