% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcgis_extras.R
\name{arc_write}
\alias{arc_write}
\title{write data to disk using arcgisbinding}
\usage{
arc_write(obj, dsn, overwrite = TRUE, ...)
}
\arguments{
\item{obj}{input source objects}

\item{dsn}{full output path}

\item{overwrite}{(optional) overwrite existing dataset. default is \code{TEUE}.}

\item{...}{(optional) other parameters passed to \code{arcgisbinding::arc.write()}}
}
\value{
An exported file in the disk.
}
\description{
write data to disk using arcgisbinding
}
\note{
The \code{IO} functions in \code{sf} and \code{terra} package can be used in most cases, and
\code{arc_write()} is not recommended unless you want to export \verb{raster layer} to
a \code{gdb} file. Because \code{arc_write()} export takeng more time.
}
