% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maptools.R
\name{geom_northarrow}
\alias{geom_northarrow}
\title{North arrow symbol}
\usage{
geom_northarrow(
  data = NULL,
  location = "tr",
  scale = 0.1,
  type = 19,
  x.min,
  x.max,
  y.min,
  y.max,
  anchor = NULL
)
}
\arguments{
\item{data}{the same \code{\link{data.frame}} passed to \code{ggplot2::ggplot()} to plot the map.}

\item{location}{string indicating the symbol's location in the plot. Possible options: "tr" (default), "br", "bl" and "tl".}

\item{scale}{number between 0 and 1 to indicate the symbol size as a proportion of the map size (bounding box).}

\item{type}{number between 1 and 19 to choose a symbol (see \code{northtypes()}).}

\item{x.min}{if \code{data} is not defined, number with the minimum x coordinate.}

\item{x.max}{if \code{data} is not defined, number with the maximum x coordinate.}

\item{y.min}{if \code{data} is not defined, number with the minimum y coordinate.}

\item{y.max}{if \code{data} is not defined, number with the maximum y coordinate.}

\item{anchor}{named \code{\link{vector}} with coordinates to control the symbol position.
For \code{location = "tr"}, \code{anchor} defines the coordinates of the symbol's topright
corner and so forth. The x coordinate must be named as x and the y coordinate as y.}
}
\description{
Adds a north symbol to maps created with ggplot2.
}
\details{
North symbols are included in the plot with the \code{ggplot2::annotation_custom()} function,
which do not works when used together with an empty call to \code{ggplot2::ggplot()}.
}
