% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maptools.R
\name{northtypes}
\alias{northtypes}
\title{Available north arrow types.}
\usage{
northtypes()
}
\description{
Displays available north arrow types.
}
\note{
The symbols were obtained from QGIS 2.8.1 - Wien.
}
\examples{
northtypes()
}
\references{
http://www.qgis.org/en/site
}
