% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcgis_extras.R
\name{rast_arc}
\alias{rast_arc}
\title{read spatraster using arcgisbinding}
\usage{
rast_arc(dsn)
}
\arguments{
\item{dsn}{The path of the source file}
}
\value{
object of class \code{SpatRaster} when a file was successfully read.
}
\description{
read spatraster using arcgisbinding
}
