% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcgis_extras.R
\name{read_arc}
\alias{read_arc}
\title{read simple features using arcgisbinding}
\usage{
read_arc(dsn)
}
\arguments{
\item{dsn}{The path of the source file}
}
\value{
object of class \code{sf} when a file was successfully read.
}
\description{
read simple features using arcgisbinding
}
