% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_extras.R
\name{st_point2polygon}
\alias{st_point2polygon}
\title{construct a polygon from a series of points}
\usage{
st_point2polygon(pcoord1, pcoord2, pcrs = NULL)
}
\arguments{
\item{pcoord1}{The x-coordinate of the points}

\item{pcoord2}{The y-coordinate of the points}

\item{pcrs}{(optional) Coordinate reference system to be assigned, default uses \code{WGS84},
more details to see \code{sf::st_crs()}}
}
\value{
An \code{sfc} polygon object.
}
\description{
construct a polygon from a series of points
}
\examples{
lon = c(126.626510,126.625261,126.626378,126.626541,126.626721,126.627732,126.626510)
lat = c(45.731596,45.729834,45.729435,45.729676,45.729604,45.730915,45.731596)
g = st_transform_cn(lon,lat)
st_point2polygon(g$lon,g$lat)
}
