% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcgis_extras.R
\name{write_arc}
\alias{write_arc}
\title{write data to disk using arcgisbinding}
\usage{
write_arc(obj, dsn, overwrite = TRUE, ...)
}
\arguments{
\item{obj}{input source objects}

\item{dsn}{full output path}

\item{overwrite}{(optional) overwrite existing dataset. default is \code{TEUE}.}

\item{...}{(optional) other parameters passed to \code{arcgisbinding::arc.write()}}
}
\value{
An exported file in the disk.
}
\description{
write data to disk using arcgisbinding
}
